Decoupled products from the shop
First implementation sucked
