+++ MiniShopify/src/main/java/org/example/ProductManagementFacade.java
@@ -1,62 +1,41 @@
 package org.example;
-/// This is a stub only here to support ShopManagementFacade and has not been fully implemented yet.
-public class ProductManagementFacade {
-    private String productName;
-    private String productDescription;
-    private double productPrice;
-    private String productCategory;
-    private int productID;
-    private int quantity;
-    private ShopManagementFacade shop;
 
-    public ProductManagementFacade(String productName, int productID, String productDescription, double productPrice, String productCategory, int quantity, ShopManagementFacade shop) {
-        this.productName = productName;
-        this.productID = productID;
-        this.productDescription = productDescription;
-        this.productPrice = productPrice;
-        this.productCategory = productCategory;
-        this.quantity = quantity;
+import java.util.ArrayList;
+
+public class ProductManagementFacade {
+
+    private ShopManagementFacade shop;
+    private ArrayList<Product> products;
+
+    public ProductManagementFacade(ShopManagementFacade shop) {
         this.shop = shop;
     }
 
-    /// For safety and simplicity there are only setters for desc. and price. Do not worry about others right now
-    public String getProductName() {
-        return productName;
+    public void createProduct (String productName, int productID, String productDescription, double productPrice, String productCategory, int quantity) {
+        products.add(new Product(productName, productID,productDescription,productPrice,productCategory, quantity, shop));
     }
 
-    public String getProductDescription() {
-        return productDescription;
+    public ArrayList<Product> getProducts() {
+        return products;
     }
 
-    public double getProductPrice() {
-        return productPrice;
+    public Product findProductByID(int productID) {
+        for (Product p : products){
+            if (p.getProductID() == productID){
+                return p;
+            }
+        }
+        System.out.println("Could not find product");
+        return null;
     }
 
-    public String getProductCategory() {
-        return productCategory;
+    public void removeProductByID(int productID) {
+        for (Product p : products){
+            if (p.getProductID() == productID) {
+                products.remove(p);
+            }
+        }
     }
+    public void removeProductByObject(Product p) {products.remove(p);}
 
-    public int getProductID() {
-        return productID;
-    }
-    public void setProductPrice(double productPrice) {
-        this.productPrice = productPrice;
-    }
-
-    public void setProductDescription(String productDescription) {
-        this.productDescription = productDescription;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public ShopManagementFacade getShop() {
-        return shop;
-    }
-
-    public void setQuantityMinus1() {
-
-        this.quantity = quantity - 1;
-    }
 }
