+++ MiniShopify/src/main/java/org/example/ShopManagementFacade.java
@@ -6,18 +6,17 @@ public class ShopManagementFacade {
 
     private String name;
     private ArrayList<String> tags;
-    private ArrayList<ProductManagementFacade> products;
     private String ShopType;
     private UserMangementFacade owner;
     private int id;
+    private ProductManagementFacade products;
+    private OrderAndCheckoutFacade orderer;
 
     public ShopManagementFacade(UserMangementFacade owner, int id) {
         this.owner = owner;
         this.id = id;
-
-
-
-
+        products = new ProductManagementFacade(this);
+        orderer = new OrderAndCheckoutFacade(this);
     }
 
     public String getName() {
@@ -40,35 +39,13 @@ public class ShopManagementFacade {
         return ShopType;
     }
 
-    public void createProduct (String productName, int productID, String productDescription, double productPrice, String productCategory, int quantity) {
-        products.add(new ProductManagementFacade(productName, productID,productDescription,productPrice,productCategory, quantity, this));
-    }
-
-    public ArrayList<ProductManagementFacade> getProducts() {
-        return products;
-    }
-
-    public ProductManagementFacade findProductByID(int productID) {
-        for (ProductManagementFacade p : products){
-            if (p.getProductID() == productID){
-                return p;
-            }
-        }
-        System.out.println("Could not find product");
-        return null;
-    }
-
-    public void removeProductByID(int productID) {
-        for (ProductManagementFacade p : products){
-            if (p.getProductID() == productID) {
-                products.remove(p);
-            }
-    }
-    }
-    public void removeProductByObject(ProductManagementFacade p) {products.remove(p);}
 
     public void setShopType(String shopType) {
         ShopType = shopType;
+    }
+
+    public ProductManagementFacade getProductManager (){
+        return products;
     }
 
     //No setters for owner and ID for now for safety reasons
